requirements.txt
# Core dependencies
mcp==1.5.0
atlassian-python-api==4.0.7
fastapi>=0.110.0
uvicorn>=0.27.0

# HTTP and API tools
requests>=2.31.0
pydantic==2.10.6

# Environment and configuration
python-dotenv==1.0.1